Converted to BigDecimal, added comment stubs
all decimal values previously in double were converted to BigDecimal to avoid precision errors. Tests changed accordingly. Also added comment stubs to all classes.

+++ src/main/java/miukkik/vacationpaycalculator/Change.java
@@ -1,12 +1,17 @@
+/**
+ * @author Mia Kallio
+ */
+
 package miukkik.vacationpaycalculator;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public class Change {
-	double amount;
+	BigDecimal amount;
 	LocalDate effectiveUntil;
 	
-	public Change (double amount) {
+	public Change (BigDecimal amount) {
 		this.amount = amount; 
 		effectiveUntil = LocalDate.MAX;
 	}
