+++ src/main/java/miukkik/vacationpaycalculator/ChangeList.java
@@ -1,5 +1,10 @@
+/**
+ * @author Mia Kallio
+ */
+
 package miukkik.vacationpaycalculator;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.ArrayList;
 
@@ -7,11 +12,11 @@ public class ChangeList {
 
 	private ArrayList<Change> changes;
 	
-	public ChangeList(double amount) {
+	public ChangeList(BigDecimal amount) {
 		changes.add(new Change(amount));
 	}
 	
-	public void addChange(double amount, LocalDate effectiveFrom) {
+	public void addChange(BigDecimal amount, LocalDate effectiveFrom) {
 		changes.get(changes.size()-1).setEffectiveUntil(effectiveFrom.minusDays(1));
 		changes.add(new Change(amount));
 	}
